requirements.txt
pyspark>=3.0.0
pandas>=1.3.0
pyyaml>=5.4.1
plotly>=5.3.0
numpy>=1.19.0
pytest>=6.2.5
